+++ lib/shark.dart
@@ -4,3 +4,4 @@ export 'package:shark/src/core/shark_core.dart';
 export 'package:shark/src/views/shark_controller.dart';
 export 'package:shark/src/views/shark_widget.dart';
 export 'package:shark/src/views/shark_bridge.dart';
+export 'package:shark/src/models/enum.dart';