+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.9
 
-* TODO: Describe initial release.
+## 1.0.0
+
+* stable official
+
+## 1.0.1
+
+* fix some problems on routing (link routing error)
+* export `SharkWidgetState` type
+* update test sample